fix: correct URL for Canon van Limburg (#1663)
* fix: correct URL for Canon van Limburg

Problem: we are only interested in the Canon van
Limburg for this source, not the entire Canon van
Nederland. However, the url pointed to
the Canon van Limburg landing page of the
Canon van Nederland site. That breaks the reverse
lookup functionality. Also broader and related
terms won't be found.

Fix: We change the url to point to the Canon van
Nederland. Also we change the source URI so it's
clear thisis just a subset of the Canon van
Nederland. We use a fragment identifier to
indicate this is the Limburg subset, following
the pattern that was used for Wikidata subsets.

Note: In the future we need to be aware that
reverse lookups will probably happen on the first
term source with the Canon van Nederland url
found. So if we add more subsets of the
Canon van Nederland, we need to take care
of that.

* Update packages/catalog/catalog/datasets/canon-van-limburg.jsonld

Co-authored-by: David de Boer <david@ddeboer.nl>

---------

Co-authored-by: Jeroen Seeverens <j.seeverens@xentropics.nl>
Co-authored-by: David de Boer <david@ddeboer.nl>